Remove circular dependency between conversation and user
